Update Tailwind executable call for v3.0
`solidus_starter_frontend` provides templates including the
`tailwindcss-rails` gem without any version constraints. So when 3.0 was
released this week, it broke CI runs for the Solidus repository.

It appears that the only breaking change is related to the `.executable`
call.

As of `tailwindcss-rails` `>= 3.0` the command for the executable was
extracted to a separate `tailwindcss-ruby` gem, which causes an error
in the rake tasks we generate for the `solidus_admin`.

The error this fixes is:

    NoMethodError: undefined method `executable' for Tailwindcss::Commands:Module

Co-authored-by: benjamin wil <benjamin@super.gd>