Update line_item spec to create a target_shipment for destroy test
This test is demonstrating a false passing because the OrderInventory
instance actually receives the `verify` message on being created, and
previous to this change, that would happen after the `expect_` line,
causing this test to pass despite `verify` not being received during
the `destory`. This change adds a `target_shipment` to the
`line_item`, which is needed to trigger the `verify` call.