Set nil tax category in line item product spec
We're seeing a spec failure related to changing setting required line
item attributes from `before_validation` to `after_initialize`. We can
"fix" the spec by setting a nil tax category, but this may not be
appropriate.

We've also added a TODO item regarding this change, as we think it may
be a breaking change requiring further investigation.

Co-authored-by: Noah Silvera <noah@super.gd>
Co-authored-by: Kendra Chateau <kendra@super.gd>